separacion de calculos de card
